requirements.txt
Flask==2.0.1
Werkzeug==2.0.1
Flask_Mail==0.9.1
mysql_connector_repackaged==0.3.1
Pillow==8.3.1
